+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/window/model/Document.java
@@ -8,6 +8,7 @@ import com.google.common.collect.ImmutableSet;
 import de.metas.document.engine.IDocument;
 import de.metas.document.engine.IDocumentBL;
 import de.metas.document.exceptions.DocumentProcessingException;
+import de.metas.i18n.BooleanWithReason;
 import de.metas.lang.SOTrx;
 import de.metas.letters.model.Letters;
 import de.metas.logging.LogManager;
@@ -32,6 +33,7 @@ import de.metas.ui.web.window.descriptor.DocumentFieldDependencyMap;
 import de.metas.ui.web.window.descriptor.DocumentFieldDependencyMap.DependencyType;
 import de.metas.ui.web.window.descriptor.DocumentFieldDescriptor;
 import de.metas.ui.web.window.descriptor.DocumentFieldWidgetType;
+import de.metas.ui.web.window.descriptor.decorator.ReadOnlyInfo;
 import de.metas.ui.web.window.exceptions.DocumentFieldNotFoundException;
 import de.metas.ui.web.window.exceptions.DocumentFieldReadonlyException;
 import de.metas.ui.web.window.exceptions.DocumentNotFoundException;
@@ -50,10 +52,13 @@ import org.adempiere.ad.expression.api.IExpression;
 import org.adempiere.ad.expression.api.IExpressionEvaluator.OnVariableNotFound;
 import org.adempiere.ad.expression.api.ILogicExpression;
 import org.adempiere.ad.expression.api.LogicExpressionResult;
+import org.adempiere.ad.expression.api.LogicExpressionResultWithReason;
 import org.adempiere.ad.ui.spi.ITabCallout;
 import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.service.ClientId;
+import org.adempiere.util.lang.ExtendedMemorizingSupplier;
 import org.adempiere.util.lang.IAutoCloseable;
+import org.adempiere.util.lang.impl.TableRecordReference;
 import org.compiere.util.Env;
 import org.slf4j.Logger;
 
@@ -73,6 +78,8 @@ import java.util.concurrent.locks.ReentrantReadWriteLock.ReadLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock.WriteLock;
 import java.util.function.IntSupplier;
 import java.util.function.Supplier;
+
+import static de.metas.ui.web.window.WindowConstants.FIELDNAME_IsSOTrx;
 
 /*
  * #%L
@@ -120,7 +127,6 @@ public final class Document
 
 	//
 	// Status
-	private boolean _new;
 	private boolean _deleted;
 	private final boolean _writable;
 	private FieldInitializationMode _initializingMode = null;
@@ -185,7 +191,6 @@ public final class Document
 		documentPath = builder.getDocumentPath();
 		windowNo = builder.getWindowNo();
 		_writable = builder.isWritable();
-		_new = builder.isNewDocument();
 		_deleted = false;
 		_staleStatus = new DocumentStaleState();
 		_lock = builder.createLock();
@@ -195,10 +200,12 @@ public final class Document
 		_validOnCheckout = DocumentValidStatus.documentInitiallyInvalid();
 		_valid = _validOnCheckout;
 
-		_saveStatusOnCheckout = DocumentSaveStatus.unknown();
+		_saveStatusOnCheckout = DocumentSaveStatus.unknown(!builder.isNewDocument());
 		_saveStatus = _saveStatusOnCheckout;
 
 		changesCollector = builder.getChangesCollector();
+
+		boolean hasNonBooleanIsSOTrx = false; // Used to make sure yes/no refList readonlyLogic doesn't default to 'N' if null
 
 		//
 		// Create document fields
@@ -215,6 +222,11 @@ public final class Document
 				{
 					Check.assumeNull(parentLinkField, "Only one parent link field shall exist but we found: {}, {}", parentLinkField, field); // shall no happen at this level
 					parentLinkField = field;
+				}
+
+				if (fieldName.equals(FIELDNAME_IsSOTrx) && !fieldDescriptor.isBooleanWidgetType())
+				{
+					hasNonBooleanIsSOTrx = true;
 				}
 			}
 			fieldsByName = fieldsBuilder.build();
@@ -252,9 +264,9 @@ public final class Document
 			if (_parentDocument == null)
 			{
 				final Optional<SOTrx> soTrx = entityDescriptor.getSOTrx();
-				if (soTrx.isPresent())
+				if (soTrx.isPresent() && !hasNonBooleanIsSOTrx)
 				{
-					setDynAttributeNoCheck("IsSOTrx", soTrx.get().isSales()); // cover the case for FieldName=IsSOTrx, DefaultValue=@IsSOTrx@
+					setDynAttributeNoCheck(FIELDNAME_IsSOTrx, soTrx.get().isSales()); // cover the case for FieldName=IsSOTrx, DefaultValue=@IsSOTrx@
 				}
 				setDynAttributeNoCheck("IsApproved", false); // cover the case for FieldName=IsApproved, DefaultValue=@IsApproved@
 			}
@@ -294,7 +306,6 @@ public final class Document
 		windowNo = from.windowNo;
 		_writable = copyMode.isWritable();
 
-		_new = from._new;
 		_deleted = from._deleted;
 		_staleStatus = new DocumentStaleState(from._staleStatus);
 		_lock = from._lock; // always share the same lock
@@ -493,7 +504,6 @@ public final class Document
 	/**
 	 * Set field's initial value
 	 *
-	 * @param documentField
 	 * @param mode               initialization mode
 	 * @param fieldValueSupplier initial value supplier
 	 */
@@ -820,8 +830,6 @@ public final class Document
 
 	/**
 	 * NOTE: API method, don't call it directly
-	 *
-	 * @param documentValuesSupplier
 	 */
 	public void refreshFromSupplier(final DocumentValuesSupplier documentValuesSupplier)
 	{
@@ -839,7 +847,6 @@ public final class Document
 				.add("tableName", entityDescriptor.getTableNameOrNull())
 				.add("parentId", parentDocument == null ? null : parentDocument.getDocumentId())
 				.add("id", getDocumentIdOrNull()) // avoid NPE
-				.add("NEW", _new ? Boolean.TRUE : null)
 				.add("windowNo", windowNo)
 				.add("writable", _writable)
 				.add("valid", _valid)
@@ -924,8 +931,6 @@ public final class Document
 	 * Sets a {@link DocumentEvaluatee} which will be used as a parent evaluatee for {@link #asEvaluatee()}.
 	 * <p>
 	 * NOTE: this shadow evaluatee is not persisted and is discarded on {@link #copy(Document, CopyMode)}.
-	 *
-	 * @param shadowParentDocumentEvaluatee
 	 */
 	public void setShadowParentDocumentEvaluatee(final IDocumentEvaluatee shadowParentDocumentEvaluatee)
 	{
@@ -998,8 +1003,7 @@ public final class Document
 
 	private IDocumentField getFieldOrNull(final String fieldName)
 	{
-		final IDocumentField documentField = fieldsByName.get(fieldName);
-		return documentField;
+		return fieldsByName.get(fieldName);
 	}
 
 	public IDocumentFieldView getFieldViewOrNull(final String fieldName)
@@ -1067,13 +1071,12 @@ public final class Document
 
 	public boolean isNew()
 	{
-		return _new;
+		return !getSaveStatus().isPresentInDatabase();
 	}
 
-	// TODO: make this method private/package
-	public void markAsNotNew()
+	public void markAsSaved()
 	{
-		_new = false;
+		setSaveStatusAndReturn(DocumentSaveStatus.saved());
 	}
 
 	/* package */ void markAsDeleted()
@@ -1125,6 +1128,11 @@ public final class Document
 		return _saveStatus;
 	}
 
+	private DocumentSaveStatus setSaveStatusAndReturn(@NonNull final Exception exception)
+	{
+		return setSaveStatusAndReturn(DocumentSaveStatus.error(exception, this._saveStatus));
+	}
+
 	private DocumentSaveStatus setSaveStatusAndReturn(@NonNull final DocumentSaveStatus saveStatus)
 	{
 		_saveStatus = saveStatus;
@@ -1161,12 +1169,12 @@ public final class Document
 			@NonNull final String fieldName,
 			@Nullable final Object value,
 			@Nullable final ReasonSupplier reason,
-			final boolean ignoreReadonlyFlag)
+			@NonNull final DocumentFieldLogicExpressionResultRevaluator readonlyRevaluator)
 			throws DocumentFieldReadonlyException
 	{
 		final IDocumentField documentField = getField(fieldName);
 
-		if (!ignoreReadonlyFlag && documentField.isReadonly())
+		if (readonlyRevaluator.isReadonly(documentField))
 		{
 			throw new DocumentFieldReadonlyException(fieldName, value);
 		}
@@ -1181,14 +1189,23 @@ public final class Document
 		}
 	}
 
-	public void processValueChanges(@NonNull final List<JSONDocumentChangedEvent> events, @Nullable final ReasonSupplier reason) throws DocumentFieldReadonlyException
+	public void processValueChanges(
+			@NonNull final List<JSONDocumentChangedEvent> events,
+			@Nullable final ReasonSupplier reason) throws DocumentFieldReadonlyException
+	{
+		processValueChanges(events, reason, DocumentFieldLogicExpressionResultRevaluator.DEFAULT);
+	}
+
+	public void processValueChanges(
+			@NonNull final List<JSONDocumentChangedEvent> events,
+			@Nullable final ReasonSupplier reason,
+			@NonNull final DocumentFieldLogicExpressionResultRevaluator readonlyRevaluator) throws DocumentFieldReadonlyException
 	{
 		for (final JSONDocumentChangedEvent event : events)
 		{
 			if (JSONDocumentChangedEvent.JSONOperation.replace == event.getOperation())
 			{
-				final boolean ignoreReadonlyFlag = false;
-				processValueChange(event.getPath(), event.getValue(), reason, ignoreReadonlyFlag);
+				processValueChange(event.getPath(), event.getValue(), reason, readonlyRevaluator);
 			}
 			else
 			{
@@ -1351,13 +1368,56 @@ public final class Document
 		getFields().forEach(documentField -> updateFieldReadOnlyAndCollect(documentField, reason));
 	}
 
+	@NonNull
 	private DocumentReadonly computeReadonly()
 	{
+		return DocumentReadonly.builder()
+				.parentActive(parentReadonly.isActive()).active(isActive())
+				.processed(parentReadonly.isProcessed() || isProcessed())
+				.processing(parentReadonly.isProcessing() || isProcessing())
+				.parentEnforcingReadOnly(parentReadonly.computeForceReadOnlyChildDocuments())
+				.fieldsReadonly(ExtendedMemorizingSupplier.of(this::computeFieldsReadOnly))
+				.build();
+	}
+
+	@NonNull
+	private ReadOnlyInfo computeFieldsReadOnly()
+	{
+		final boolean isReadOnlyLogicTrue = computeDefaultFieldsReadOnly().booleanValue();
+		if (isReadOnlyLogicTrue)
+		{
+			return ReadOnlyInfo.TRUE;
+		}
+
+		if (isFieldsReadOnlyInUI())
+		{
+			return ReadOnlyInfo.TRUE;
+		}
+
+		final TableRecordReference recordReference = this.getTableRecordReference().orElse(null);
+		if (recordReference == null)
+		{
+			return ReadOnlyInfo.of(BooleanWithReason.FALSE);
+		}
+
+		return getEntityDescriptor()
+				.getDocumentDecorators()
+				.stream()
+				.map(documentDecorator -> documentDecorator.isReadOnly(recordReference))
+				.filter(ReadOnlyInfo::isReadOnly)
+				.findFirst()
+				.orElse(ReadOnlyInfo.FALSE);
+	}
+
+	@NonNull
+	private LogicExpressionResult computeDefaultFieldsReadOnly()
+	{
 		final ILogicExpression allFieldsReadonlyLogic = getEntityDescriptor().getReadonlyLogic();
-		LogicExpressionResult allFieldsReadonly;
+
+		final LogicExpressionResult allFieldsReadonly;
 		try
 		{
-			allFieldsReadonly = allFieldsReadonlyLogic.evaluateToResult(asEvaluatee(), OnVariableNotFound.Fail);
+			return allFieldsReadonlyLogic.evaluateToResult(asEvaluatee(), OnVariableNotFound.Fail);
 		}
 		catch (final Exception e)
 		{
@@ -1365,13 +1425,7 @@ public final class Document
 			logger.warn("Failed evaluating entity readonly logic {} for {}. Considering {}", allFieldsReadonlyLogic, this, allFieldsReadonly, e);
 		}
 
-		final DocumentReadonly readonlyComputed = DocumentReadonly.builder()
-				.parentActive(parentReadonly.isActive()).active(isActive())
-				.processed(parentReadonly.isProcessed() || isProcessed())
-				.processing(parentReadonly.isProcessing() || isProcessing())
-				.fieldsReadonly(allFieldsReadonly.booleanValue())
-				.build();
-		return readonlyComputed;
+		return allFieldsReadonly;
 	}
 
 	private void updateFieldReadOnlyAndCollect(final IDocumentField documentField, final ReasonSupplier reason)
@@ -1390,8 +1444,15 @@ public final class Document
 	{
 		// Check document's readonly logic
 		final DocumentReadonly documentReadonlyLogic = getReadonly();
-		if (documentReadonlyLogic.computeFieldReadonly(documentField.getFieldName(), documentField.isAlwaysUpdateable()))
+		final BooleanWithReason isReadOnly = documentReadonlyLogic.computeFieldReadonly(documentField.getFieldName(), documentField.isAlwaysUpdateable());
+
+		if (isReadOnly.isTrue())
 		{
+			if (WindowConstants.FIELDNAME_DocumentSummary.equals(documentField.getFieldName()))
+			{
+				return new LogicExpressionResultWithReason(LogicExpressionResult.TRUE, isReadOnly.getReason());
+			}
+
 			return LogicExpressionResult.TRUE;
 		}
 
@@ -1609,6 +1670,12 @@ public final class Document
 
 	/* package */ void deleteIncludedDocuments(final DetailId detailId, final DocumentIdsSelection rowIds)
 	{
+		final BooleanWithReason isDeleteSubDocumentsForbidden = isDeleteSubDocumentsForbidden();
+		if (isDeleteSubDocumentsForbidden.isTrue())
+		{
+			throw new AdempiereException(isDeleteSubDocumentsForbidden.getReason()).markAsUserValidationError();
+		}
+
 		final IIncludedDocumentsCollection includedDocuments = getIncludedDocumentsCollection(detailId);
 		includedDocuments.deleteDocuments(rowIds);
 		checkAndGetValidStatus();
@@ -1637,6 +1704,12 @@ public final class Document
 		return isActiveField == null || isActiveField.getValueAsBoolean(); // active if field not found (shall not happen)
 	}
 
+	/* package */ boolean isFieldsReadOnlyInUI()
+	{
+		final IDocumentFieldView isFieldsReadOnlyInUI = getFieldUpToRootOrNull(WindowConstants.FIELDNAME_IsFieldsReadOnlyInUI);
+		return isFieldsReadOnlyInUI != null && isFieldsReadOnlyInUI.getValueAsBoolean();
+	}
+
 	/* package */ void setParentReadonly(@NonNull final DocumentReadonly parentReadonly)
 	{
 		final DocumentReadonly parentReadonlyOld = this.parentReadonly;
@@ -1652,9 +1725,6 @@ public final class Document
 	/**
 	 * Set Dynamic Attribute.
 	 * A dynamic attribute is an attribute that is not stored in database and is kept as long as this this instance is not destroyed.
-	 *
-	 * @param name
-	 * @param value
 	 */
 	public Object setDynAttribute(final String name, final Object value)
 	{
@@ -1679,7 +1749,6 @@ public final class Document
 	/**
 	 * Get Dynamic Attribute
 	 *
-	 * @param name
 	 * @return attribute value or null if not found
 	 */
 	public <T> T getDynAttribute(final String name)
@@ -1691,8 +1760,6 @@ public final class Document
 	/**
 	 * Get Dynamic Attribute
 	 *
-	 * @param name
-	 * @param defaultValue
 	 * @return attribute value or <code>defaultValue</code> if not found
 	 */
 	public <T> T getDynAttribute(final String name, final T defaultValue)
@@ -1736,7 +1803,7 @@ public final class Document
 		};
 
 		OnValidStatusChanged MARK_NOT_SAVED = (document, invalidStatus) -> {
-			document.setSaveStatusAndReturn(DocumentSaveStatus.notSaved(invalidStatus));
+			document.setSaveStatusAndReturn(DocumentSaveStatus.notSaved(invalidStatus, document.getSaveStatus()));
 		};
 
 	}
@@ -1880,7 +1947,7 @@ public final class Document
 		if (!validState.isValid())
 		{
 			logger.debug("Skip saving because document {} is not valid: {}", this, validState);
-			return setSaveStatusAndReturn(DocumentSaveStatus.notSaved(validState));
+			return setSaveStatusAndReturn(DocumentSaveStatus.notSaved(validState, getSaveStatus()));
 		}
 
 		//
@@ -1894,7 +1961,7 @@ public final class Document
 			// NOTE: usually if we do the right checks we shall not get to this
 			// logger.warn("Failed saving document, but IGNORED: {}", this, saveEx);
 			setValidStatusAndReturn(DocumentValidStatus.invalid(saveEx), OnValidStatusChanged.DO_NOTHING);
-			return setSaveStatusAndReturn(DocumentSaveStatus.error(saveEx));
+			return setSaveStatusAndReturn(saveEx);
 		}
 	}
 
@@ -1920,7 +1987,7 @@ public final class Document
 			}
 			catch (Exception ex)
 			{
-				return setSaveStatusAndReturn(DocumentSaveStatus.error(ex)).throwIfError();
+				return setSaveStatusAndReturn(ex).throwIfError();
 			}
 		}
 		else
@@ -2059,6 +2126,61 @@ public final class Document
 		}
 
 		return standardActions;
+	}
+
+	@NonNull
+	public Optional<TableRecordReference> getTableRecordReference()
+	{
+		final String tableName = entityDescriptor.getTableNameOrNull();
+		final Integer recordId = getDocumentId().isInt() ? getDocumentIdAsInt() : null;
+
+		if (tableName == null || recordId == null)
+		{
+			return Optional.empty();
+		}
+
+		return Optional.of(TableRecordReference.of(tableName, recordId));
+	}
+
+	@NonNull
+	public BooleanWithReason isDeleteForbidden()
+	{
+		final TableRecordReference recordReference = getTableRecordReference().orElse(null);
+
+		if (recordReference == null)
+		{
+			return BooleanWithReason.FALSE;
+		}
+
+		return entityDescriptor.getDocumentDecorators()
+				.stream()
+				.map(decorator -> decorator.isDeleteForbidden(recordReference))
+				.filter(BooleanWithReason::isTrue)
+				.findFirst()
+				.orElse(BooleanWithReason.FALSE);
+	}
+
+	public boolean containsField(@NonNull final String fieldName)
+	{
+		return fieldsByName.containsKey(fieldName);
+	}
+
+	@NonNull
+	private BooleanWithReason isDeleteSubDocumentsForbidden()
+	{
+		final TableRecordReference recordReference = getTableRecordReference().orElse(null);
+
+		if (recordReference == null)
+		{
+			return BooleanWithReason.FALSE;
+		}
+
+		return entityDescriptor.getDocumentDecorators()
+				.stream()
+				.map(decorator -> decorator.isDeleteSubDocumentsForbidden(recordReference))
+				.filter(BooleanWithReason::isTrue)
+				.findFirst()
+				.orElse(BooleanWithReason.FALSE);
 	}
 
 	//
